+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0
+
+* Add `rollback` and `reset` api to rollback or reset migrations, so they can be executed again.
+
+## 1.0.1
+
+* Refactor how migrations are stored, increasing performance slightly
+
 ## 1.0.0
 
 * Initial release with basic only once execution of migrations logic